requirements.txt
python-telegram-bot==22.5
pillow==12.1.0